Update runtime and sources to latest version.
Runtime v40 was deprecated recently, so this bumps it to v42.
The launcher and the game seems to work fine with it in my tests.

The game URL has changed from a mirror to an upstream URL.
And the launcher source had to be changed from Kekun's to mine as
the appdata is tied alongside it.